+++ tests/test_p_less.py
@@ -74,9 +74,10 @@ def test_less_render():
         remove_file(filepath)
 
 
-def test_less_make():
+def test_less_build():
     filepath = make_view(FILENAME_IN, SRC_LESS)
     filepath_out = get_build_filepath(FILENAME_OUT)
+    clay_.settings['FILTER_PARTIALS'] = False
     try:
         clay_.build()
         content = read_file(filepath_out).strip()