+++ tests/test_p_markdown.py
@@ -65,6 +65,7 @@ def test_markdown_render():
 def test_markdown_build():
     filepath = make_view(FILENAME_IN, SRC_MARKDOWN)
     filepath_out = get_build_filepath(FILENAME_OUT)
+    clay_.settings['FILTER_PARTIALS'] = False
     try:
         clay_.build()
         content = read_file(filepath_out).strip()