+++ CHANGES.md
@@ -1,13 +1,21 @@
 # Clay Changelog
 
 
+## Version 1.4
+
+- New settings:
+    * FILTER_PARTIALS: Do not show HTML fragments in the final build
+    * VIEWS_INCLUDE: When building, include all views listed here, even if it's a HTML fragment and FILTER_PARTIALS is True
+
+
+
 ## Version 1.3
 
 - Fix nasty bug that in some cases prevent it from running when some extra plugin libraries where missing.
 
 - All settings names are upper cased by default (altough lower case versions are supported as well).
 
-- Default jQuery updated to 1.8.1
+- Default jQuery updated to 1.8.2
 
 - Merged a README typo fix by @z4y4ts (thanks).